+++ packages/hull-vm/src/ui/index.js
@@ -13,3 +13,4 @@ export { default as RecentEntriesUI } from "./recent-ui";
 export { default as VirtualMachineUI } from "./vm-ui";
 export { default as ProcessorUI } from "./processor-ui";
 export { default as ReplUI } from "./repl-ui";
+export { default as OutgoingUI } from "./outgoing-ui";